refactor(jest-snapshot): refactor Prettier related part
