+++ packages/jest-snapshot/package.json
@@ -37,9 +37,9 @@
     "jest-matcher-utils": "workspace:*",
     "jest-message-util": "workspace:*",
     "jest-util": "workspace:*",
+    "make-synchronized": "^0.3.1",
     "pretty-format": "workspace:*",
-    "semver": "^7.5.3",
-    "synckit": "^0.9.0"
+    "semver": "^7.5.3"
   },
   "devDependencies": {
     "@babel/preset-flow": "^7.7.2",
